Rename to MultipleUploadFilePickerTab
