+++ src/controls/filePicker/MultipleUploadFilePickerTab/IMultipleUploadFilePickerTabProps.ts
@@ -1,6 +1,6 @@
 import { IFilePickerResult, IFilePickerTab } from "../FilePicker.types";
 
-export interface IMultipleUploadFilePicketTabProps extends IFilePickerTab {
+export interface IMultipleUploadFilePickerTabProps extends IFilePickerTab {
   onChange: (value: IFilePickerResult) => void;
   renderCustomMultipleUploadTabContent: (filePickerResult: IFilePickerResult) => JSX.Element | null;
 }
