+++ src/controls/filePicker/MultipleUploadFilePickerTab/index.ts
@@ -1,0 +1,3 @@
+export * from './MultipleUploadFilePickerTab';
+export * from './IMultipleUploadFilePickerTabProps';
+export * from './IMultipleUploadFilePickerTabState';
