--- src/controls/filePicker/MultipleUploadFilePicketTab/index.ts
@@ -1,3 +1,0 @@
-export * from './MultipleUploadFilePicketTab';
-export * from './IMultipleUploadFilePicketTabProps';
-export * from './IMultipleUploadFilePicketTabState';
